requirements.txt
numpy==1.16.4
pandas==0.25.3
scikit-learn==0.21.3
scipy==1.3.3
six==1.14.0
tensorflow-gpu==1.14.0
keras==2.3.1
networkx==2.4